[FIX] account: Access cash rounding
In order to see the fields profit_account_id and loss_account_id, we had to be part of the group_account_user which is not available when the module accounting is not used.
Now, to be able to use it when invoicing, we change the group to the group_account_invoice.

closes odoo/odoo#41310

X-original-commit: a65c5cea2f5c2c6d0b99e95fdd29b6a286200723
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>