[FIX] point_of_sale: add UoM to invoice line
Enable units of measure from Settings.
Go to Point of Sale, enable "Invoicing" from the PoS, open the session,
sell something specifing customer and invoice. Close the session and
post.
Check the sale order, UoM is specified. Now check the associated
invoice, UoM is missing.

This is because the UoM is not propagated from the sale order to the
invoice. Adding the missing value

opw-2129428

closes odoo/odoo#41040

X-original-commit: 2d5f22f040e7d28a17191cd6cc512b58b929efa1
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>