[IMP] project: minimal chatter on project form view
Before this commit, chatter of project form view took about 1/3rd
of screen width on big screen, even though no discussion is possible.

This commit solves this issue by moving the chatter at the bottom of
the sheet, so that it isn't displayed at the right of the screen.

Task-2125078

closes odoo/odoo#41082

Related: odoo/enterprise#6959
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>