[FIX] base: ir.ui.menu, allow unlink of menu_item with unactive children
- Set active = false on menu that has a parent
- Delete the parent menu.

It crashes in the unlink method of ir_ui_menu
because it didn't set parent=null on records that are inactive.

Because the search method to get children
of a menu item only returns node where active=true
if `active_test` is not passed in the context to
disable this active test filter.

closes odoo/odoo#41859

X-original-commit: a9c0c1d45583d547754735e857168d5ae27876a5
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>