[REV] mail: prevent message loss during connection issue
This reverts commit 43d5dc2421a33b35aacd36e4861d8ebd81dd0114.

This commit prevented an issue of losing messages on post in case
of connectivity loss. The implementation consists of keeping the
content in the composer until the server acknowledges handling the
message.

However, this solution introduces significant issues. For instance,
sending messages in quick succession will likely loose or duplicate
some content, based on how fast/slow the RPC `message_post` is
resolved. Plus, there was an issue on mobile because the input focus
was lost during the RPC. As a consequence, the virtual keyboard
flickered.

The intent was to fix a rare issue, and since the problems of this
solution clearly outweigh the gain, we decided to revert.

closes odoo/odoo#41544

X-original-commit: 36b2e99e08bfc6d92a821d0346e783695580abf9
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>